리뷰 관련 오류 수정

+++ VamPa/src/main/java/com/vam/model/UpdateReplyDTO.java
@@ -4,7 +4,7 @@ public class UpdateReplyDTO {
 
 	private int bookId;
 	
-	private int ratingAvg;
+	private double ratingAvg;
 
 	public int getBookId() {
 		return bookId;
@@ -14,11 +14,11 @@ public class UpdateReplyDTO {
 		this.bookId = bookId;
 	}
 
-	public int getRatingAvg() {
+	public double getRatingAvg() {
 		return ratingAvg;
 	}
 
-	public void setRatingAvg(int ratingAvg) {
+	public void setRatingAvg(double ratingAvg) {
 		this.ratingAvg = ratingAvg;
 	}
 
@@ -26,7 +26,8 @@ public class UpdateReplyDTO {
 	public String toString() {
 		return "UpdateReplyDTO [bookId=" + bookId + ", ratingAvg=" + ratingAvg + "]";
 	}
-	
+
+
 	
 	
 }
