+++ VamPa_MySQL/src/main/java/com/vam/model/UpdateReplyDTO.java
@@ -4,7 +4,7 @@ public class UpdateReplyDTO {
 
 	private int bookId;
 	
-	private int ratingAvg;
+	private double ratingAvg;
 
 	public int getBookId() {
 		return bookId;
@@ -14,17 +14,17 @@ public class UpdateReplyDTO {
 		this.bookId = bookId;
 	}
 
-	public int getRatingAvg() {
+	public double getRatingAvg() {
 		return ratingAvg;
 	}
 
-	public void setRatingAvg(int ratingAvg) {
+	public void setRatingAvg(double ratingAvg) {
 		this.ratingAvg = ratingAvg;
 	}
 
 	@Override
 	public String toString() {
 		return "UpdateReplyDTO [bookId=" + bookId + ", ratingAvg=" + ratingAvg + "]";
-	}	
+	}
 	
 }
